Added RedirectAlowedUrls
