+++ framework/src/Volo.Abp.UI.Navigation/Volo/Abp/Ui/Navigation/Urls/AppUrlOptions.cs
@@ -1,12 +1,17 @@
-﻿namespace Volo.Abp.UI.Navigation.Urls
+﻿using System.Collections.Generic;
+
+namespace Volo.Abp.UI.Navigation.Urls
 {
     public class AppUrlOptions
     {
         public ApplicationUrlDictionary Applications { get; }
 
+        public List<string> RedirectAllowedUrls { get; }
+
         public AppUrlOptions()
         {
             Applications = new ApplicationUrlDictionary();
+            RedirectAllowedUrls = new List<string>();
         }
     }
 }
