+++ framework/src/Volo.Abp.UI.Navigation/Volo/Abp/Ui/Navigation/Urls/IAppUrlProvider.cs
@@ -6,5 +6,7 @@ namespace Volo.Abp.UI.Navigation.Urls
     public interface IAppUrlProvider
     {
         Task<string> GetUrlAsync([NotNull] string appName, [CanBeNull] string urlName = null);
+
+        bool IsRedirectAllowedUrl(string url);
     }
 }
