+++ modules/account/src/Volo.Abp.Account.Web/Pages/Account/AccountPageModel.cs
@@ -25,11 +25,6 @@ namespace Volo.Abp.Account.Web.Pages.Account
             ObjectMapperContext = typeof(AbpAccountWebModule);
         }
 
-        protected virtual RedirectResult RedirectSafely(string returnUrl, string returnUrlHash = null)
-        {
-            return Redirect(GetRedirectUrl(returnUrl, returnUrlHash));
-        }
-
         protected virtual void CheckIdentityErrors(IdentityResult identityResult)
         {
             if (!identityResult.Succeeded)
@@ -40,44 +35,12 @@ namespace Volo.Abp.Account.Web.Pages.Account
             //identityResult.CheckErrors(LocalizationManager); //TODO: Get from old Abp
         }
 
-        protected virtual string GetRedirectUrl(string returnUrl, string returnUrlHash = null)
-        {
-            returnUrl = NormalizeReturnUrl(returnUrl);
-
-            if (!returnUrlHash.IsNullOrWhiteSpace())
-            {
-                returnUrl = returnUrl + returnUrlHash;
-            }
-
-            return returnUrl;
-        }
-
-        protected virtual string NormalizeReturnUrl(string returnUrl)
-        {
-            if (returnUrl.IsNullOrEmpty())
-            {
-                return GetAppHomeUrl();
-            }
-
-            if (Url.IsLocalUrl(returnUrl))
-            {
-                return returnUrl;
-            }
-
-            return GetAppHomeUrl();
-        }
-
         protected virtual void CheckCurrentTenant(Guid? tenantId)
         {
             if (CurrentTenant.Id != tenantId)
             {
                 throw new ApplicationException($"Current tenant is different than given tenant. CurrentTenant.Id: {CurrentTenant.Id}, given tenantId: {tenantId}");
             }
-        }
-
-        protected virtual string GetAppHomeUrl()
-        {
-            return "~/"; //TODO: ???
         }
     }
 }
